petfriend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.0
Django==4.1.10
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-phonenumber-field==7.1.0
django-rest-framework==0.1.0
django-timezone-field==6.0
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.3.1
flake8==6.1.0
future==0.18.3
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.1
mccabe==0.7.0
packaging==23.1
phonenumberslite==8.13.11
Pillow==9.5.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2023.3
rabbitmq==0.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
